time.h: struct tm tm_gmtoff is long, fixing tm_zone offset
`tm_gmtoff` was declared `int`, but on macOS and glibc it is a `long`
(8 bytes, 8-aligned). The narrower field placed the following `tm_zone`
pointer at struct offset 40 instead of 48, so after localtime() filled
the host layout the program read tm_zone from the wrong offset and got a
stray pointer. CPython's time-module timezone setup dereferenced it in
strncpy and crashed. The members are now byte-exact with the host
(sizeof 56, tm_gmtoff at 40, tm_zone at 48).

Regression: struct_tm_tm_zone_offset.c (checks the offsets and that a
localtime-filled tm_zone is a readable string).

The python demo's smoke check now sets PYTHONPATH to the source tree's
Lib/ so the interpreter finds the encodings module (the reference build
needs it too), and the harness decodes subprocess output with
errors="replace" so non-UTF-8 test output doesn't abort it.

+++ headers/include/time.h
@@ -73,11 +73,11 @@ struct timeval {
 };
 #endif
 
-// `struct tm` exposed as an opaque mirror of the libc shape so
-// programs that walk the broken-down components (`tm_year`,
-// `tm_mon`, ...) typecheck. Layout matches the common Unix /
-// glibc / musl encoding; callers that hand pointers to libc
-// don't need it to be byte-exact, just large enough.
+// `struct tm`. The host's localtime / gmtime fill this and the program
+// reads the fields back, so the BSD extension members must sit at the
+// host's offsets. On macOS and glibc `tm_gmtoff` is a `long` (8 bytes,
+// 8-aligned), which places `tm_zone` at offset 48; an `int tm_gmtoff`
+// would put `tm_zone` at offset 40 and read a garbage pointer.
 struct tm {
     int tm_sec;
     int tm_min;
@@ -88,7 +88,7 @@ struct tm {
     int tm_wday;
     int tm_yday;
     int tm_isdst;
-    int tm_gmtoff;
+    long tm_gmtoff;
     char *tm_zone;
 };
